backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.10"
dependencies = [
    "dj-database-url>=3.1.0",
    "django>=5.2.11",
    "django-cors-headers>=4.9.0",
    "djangorestframework>=3.16.1",
    "psycopg[binary]>=3.3.2",
    "pyjwt[crypto]>=2.11.0",
]
